+++ internal/asm/dwcas_fallback.go
@@ -6,6 +6,6 @@ import (
 	"unsafe"
 )
 
-func DWCAS(ptr unsafe.Pointer, old1 uintptr, old2 unsafe.Pointer, new1 uintptr, new2 unsafe.Pointer) bool {
+func DWCAS(ptr unsafe.Pointer, old1 uint64, old2 unsafe.Pointer, new1 uint64, new2 unsafe.Pointer) bool {
 	panic("cc: DWCAS fallback not implemented")
 }
